Revert "Keyguard: Create lockscreen weather, move weatherimpl to utils."
This reverts commit d855e463bb934f41cee72a893609626de0a03f25.
