+++ packages/Keyguard/AndroidManifest.xml
@@ -40,7 +40,6 @@
     <uses-permission android:name="android.permission.MEDIA_CONTENT_CONTROL" />
     <uses-permission android:name="android.permission.ACCESS_KEYGUARD_SECURE_STORAGE" />
     <uses-permission android:name="android.permission.TRUST_LISTENER" />
-    <uses-permission android:name="com.cyanogenmod.lockclock.permission.READ_WEATHER" />
 
     <application android:label="@string/app_name"
         android:process="com.android.systemui"
