--- packages/Keyguard/res/layout/keyguard_weather_view.xml
@@ -1,56 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!--
-**
-** Copyright 2015, The CyanogenMod Project
-**
-** Licensed under the Apache License, Version 2.0 (the "License")
-** you may not use this file except in compliance with the License.
-** You may obtain a copy of the License at
-**
-**     http://www.apache.org/licenses/LICENSE-2.0
-**
-** Unless required by applicable law or agreed to in writing, software
-** distributed under the License is distributed on an "AS IS" BASIS,
-** WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-** See the License for the specific language governing permissions and
-** limitations under the License.
-*/
--->
-
-<!-- This is a view that shows general weather information in Keyguard. -->
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-        android:id="@+id/keyguard_weather_view"
-        android:layout_centerHorizontal="true"
-        android:layout_gravity="bottom"
-        android:paddingBottom="14dp"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:orientation="vertical"   >
-
-    <ImageView
-        android:id="@+id/weather_image"
-        android:scaleType="centerCrop"
-        android:layout_width="wrap_content"
-        android:layout_marginBottom="-10dp"
-        android:layout_height="wrap_content" />
-
-    <TextView
-        android:textColor="@color/clock_white"
-        android:textSize="16sp"
-        android:layout_gravity="center_horizontal"
-        android:includeFontPadding="false"
-        android:id="@+id/temperature"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content" />
-
-    <TextView
-        android:textColor="@color/clock_white"
-        android:textSize="16sp"
-        android:layout_gravity="center_horizontal"
-        android:includeFontPadding="false"
-        android:visibility="gone"
-        android:id="@+id/city"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content" />
-
-</LinearLayout>
