+++ packages/SystemUI/src/com/android/systemui/statusbar/policy/WeatherController.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2015 The CyanogenMod Project
+ * Copyright (C) 2014 The CyanogenMod Project
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,9 +14,7 @@
  * limitations under the License.
  */
 
-package com.android.internal.util.cm;
-
-import android.graphics.drawable.Drawable;
+package com.android.systemui.statusbar.policy;
 
 public interface WeatherController {
     void addCallback(Callback callback);
@@ -30,7 +28,5 @@ public interface WeatherController {
         public String temp = null;
         public String city = null;
         public String condition = null;
-        public int conditionCode = 0;
-        public Drawable conditionDrawable = null;
     }
 }
